feat: implement adaptive window size classes and OAuth handling
- Added adaptive window size handling for better UI responsiveness across different devices.
- Refactored OAuth flow handling by introducing a dedicated callback mechanism in `MainActivity`.
- Enhanced logging setup using `Napier` with updated parameter ordering.
- Integrated new `SubscriptionRepository` for loading subscriptions, replacing direct API calls.
- Updated dependencies in `build.gradle.kts` and `libs.versions.toml` with the latest versions for adaptive and material3 libraries.

> GENERATE BY https://aicommit.app

+++ composeApp/src/commonMain/kotlin/me/rosuh/data/model/PostEntriesRequest.kt
@@ -101,8 +101,8 @@ data class User(
 data class Media(
     val type: String,
     val url: String,
-    val width: Int? = null,
-    val height: Int? = null,
+    val width: Double? = null,
+    val height: Double? = null,
     @SerialName("preview_image_url") val previewImageUrl: String? = null,
     val blurhash: String? = null
 )
